LXDV-76 (#41)
* npm install TS version for lint

* fixed all Lint warning
